requirements.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.11.2
cramjam==2.6.2
Django==4.1.7
django-bootstrap-v5==1.0.11
django-filter==22.1
django-tables2==2.5.3
fastparquet==2023.2.0
fsspec==2023.3.0
nfl-data-py==0.3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.7.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
